requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.4
djangorestframework==3.15.1
psycopg2==2.9.9
sqlparse==0.5.0
typing_extensions==4.11.0
